feat(loader): add MustLoadX function for Loader
- Add MustLoadX method to Core struct
- Update Application to use MustLoadX instead of Loads method- Add unit tests for MustLoadX
- Refactor Load and MustLoad methods to use the new MustLoadX implementation
