+++ core.go
@@ -106,128 +106,6 @@ type Core struct {
 	log                Logger `gone:"*"`
 }
 
-// Load loads a Goner into the Core with optional configuration options.
-//
-// A Goner can be registered by name if it implements NamedGoner interface.
-// If a Goner with the same name already exists:
-// - If forceReplace option is true, the old one will be replaced
-// - Otherwise returns LoadedError
-//
-// If the Goner implements Provider interface (has Provide method), it will be registered as a provider.
-// If a provider for the same type already exists:
-// - If forceReplace option is true, the old one will be replaced
-// - Otherwise returns LoadedError
-//
-// Parameters:
-//   - goner: The Goner instance to load
-//   - options: Optional configuration options for the Goner
-//
-// Available Options:
-//   - Name(name string): Set custom name for the Goner
-//   - IsDefault(): Mark this Goner as the default implementation
-//   - OnlyForName(): Only register by name, not as provider
-//   - ForceReplace(): Replace existing Goner with same name/type
-//   - Order(order int): Set initialization order (lower runs first)
-//   - FillWhenInit(): Fill dependencies during initialization
-//
-// Returns error if:
-//   - Any option.Apply() fails
-//   - A Goner with same name already exists (without forceReplace)
-//   - A Provider for same type already exists (without forceReplace)
-func (s *Core) Load(goner Goner, options ...Option) error {
-	if goner == nil {
-		return NewInnerError("goner cannot be nil - must provide a valid Goner instance", LoadedError)
-	}
-	co := newCoffin(goner)
-
-	if namedGoner, ok := goner.(NamedGoner); ok {
-		co.name = namedGoner.GonerName()
-	}
-
-	for _, option := range options {
-		if err := option.Apply(co); err != nil {
-			return ToError(err)
-		}
-	}
-
-	if co.name != "" {
-		if oldCo, ok := s.nameMap[co.name]; ok {
-			if co.forceReplace {
-				for i := range s.coffins {
-					if s.coffins[i] == oldCo {
-						s.coffins[i] = co
-					}
-				}
-				s.nameMap[co.name] = co
-			} else {
-				return NewInnerErrorWithParams(LoadedError, "goner with name %q is already loaded - use ForceReplace() option to override", co.name)
-			}
-		} else {
-			s.nameMap[co.name] = co
-		}
-	}
-
-	var forceReplaceFind = false
-	if co.forceReplace {
-		for i := range s.coffins {
-			if s.coffins[i] == co {
-				s.coffins[i] = co
-				forceReplaceFind = true
-				break
-			}
-		}
-	}
-
-	if !forceReplaceFind {
-		s.coffins = append(s.coffins, co)
-	}
-
-	if co.provider != nil {
-		provider := co.provider
-
-		if co.onlyForName {
-			return nil
-		}
-
-		if oldCo, ok := s.typeProviderDepMap[provider.Type()]; ok {
-			if oldCo.goner == goner {
-				return NewInnerErrorWithParams(LoadedError, "provider for type %s is already registered with the same goner instance", GetTypeName(provider.Type()))
-			}
-
-			if co.forceReplace {
-				for i := range s.coffins {
-					if s.coffins[i] == oldCo {
-						s.coffins[i] = co
-					}
-				}
-				s.typeProviderDepMap[provider.Type()] = co
-				s.typeProviderMap[provider.Type()] = provider
-			} else {
-				return NewInnerErrorWithParams(LoadedError, "provider for type %s is already registered - use ForceReplace() option to override", GetTypeName(provider.Type()))
-			}
-		} else {
-			s.typeProviderMap[provider.Type()] = provider
-			s.typeProviderDepMap[provider.Type()] = co
-		}
-	}
-	if provider, ok := co.goner.(NamedProvider); ok {
-		for t := range co.defaultTypeMap {
-			if _, ok := s.typeProviderDepMap[t]; ok {
-				return NewInnerErrorWithParams(LoadedError, "provider for type %s is already registered - cannot use IsDefault option when Loading named provider: %T(name=%s)", GetTypeName(t), provider, provider.GonerName())
-			}
-			s.typeProviderDepMap[t] = co
-		}
-	}
-	return nil
-}
-
-func (s *Core) MustLoad(goner Goner, options ...Option) Loader {
-	if err := s.Load(goner, options...); err != nil {
-		panic(err)
-	}
-	return s
-}
-
 // Check performs dependency validation and determines initialization order:
 // 1. Collects all dependencies between components
 // 2. Validates there are no circular dependencies
@@ -670,13 +548,4 @@ func (s *Core) InjectWrapFunc(fn any, injectBefore FuncInjectHook, injectAfter F
 		}
 		return results
 	}, nil
-}
-
-func (s *Core) Loaded(key LoaderKey) bool {
-	if _, ok := s.loaderMap[key]; ok {
-		return true
-	} else {
-		s.loaderMap[key] = true
-		return false
-	}
 }
